requirements.txt
langchain[google-genai]
google-generativeai
edge-tts
asyncio-mqtt
pydub
python-dotenv
rich
click
aiofiles
sounddevice
numpy
fastapi
uvicorn
websockets
grpcio
grpcio-tools
protobuf